requirements.txt
python-dotenv
pydantic-ai-slim[google]
python-telegram-bot